client almost builds

+++ src/EterLib/Pool.h
@@ -246,7 +246,7 @@ class CPooledObject
 		{
 		}
 
-        void * operator new(unsigned int /*mem_size*/)
+        void * operator new(size_t /*mem_size*/)
         {
             return ms_kPool.Alloc();
         }
