+++ src/ScriptLib/PythonLauncher.cpp
@@ -1,6 +1,6 @@
 #include "StdAfx.h"
-#include <Python-2.7/frameobject.h>
-#include "../eterPack/EterPackManager.h"
+#include <python/frameobject.h>
+#include "eterPack/EterPackManager.h"
 
 #include "PythonLauncher.h"
 
@@ -92,7 +92,7 @@ int TraceFunc(PyObject * obj, PyFrameObject * f, int what, PyObject *arg)
 			PyTuple_GetObject(arg, 1, &exc_value);
 			PyTuple_GetObject(arg, 2, &exc_traceback);
 
-			int len;
+			Py_ssize_t len;
 			const char * exc_str;
 			PyObject_AsCharBuffer(exc_type, &exc_str, &len);
